Fix test assertion for ProfitYear type casting
Updated the test in AccountHistoryReportServiceTests to cast the `ProfitYear` property to an integer before asserting it is greater than 0. This ensures proper type handling during the comparison and resolves potential issues with type mismatches.

+++ src/services/tests/Demoulas.ProfitSharing.UnitTests/Reports/Adhoc/AccountHistoryReportServiceTests.cs
@@ -312,7 +312,7 @@ public class AccountHistoryReportServiceTests : ApiTestBase<Api.Program>
             record.BadgeNumber.ShouldBe(badgeNumber);
             record.FullName.ShouldNotBeNullOrEmpty();
             record.Ssn.ShouldNotBeNullOrEmpty();
-            record.ProfitYear.ShouldBeGreaterThan(0);
+            ((int)record.ProfitYear).ShouldBeGreaterThan(0);
             record.Contributions.ShouldBeGreaterThanOrEqualTo(0);
             record.Earnings.ShouldBeGreaterThanOrEqualTo(0);
             record.Forfeitures.ShouldBeGreaterThanOrEqualTo(0);
